+++ src/RdfaLiteMicrodata/Ports/Parser/Microdata.php
@@ -58,9 +58,9 @@ class Microdata extends AbstractParser
      * @param callable|null $errorHandler Custom HTML parsing error handler
      * @return \stdClass Extracted things
      */
-    public function parseHtmlFile($file, callable $errorHandler = null)
+    public function parseHtmlFile($file, array $errorHandler = null)
     {
-        return $this->parseHtml($this->getFileContents($file), $errorHandler);
+        return $this->parseHtml($this->getFileContents($file));
     }
 
     /**
@@ -70,11 +70,11 @@ class Microdata extends AbstractParser
      * @param callable|null $errorHandler Custom HTML parsing error handler
      * @return \stdClass Extracted things
      */
-    public function parseHtml($string, callable $errorHandler = null)
+    public function parseHtml($string, array $errorHandler = null)
     {
         return $this->parseSource(
             $string,
-            new HtmlDocumentFactory($errorHandler),
+            new HtmlDocumentFactory(),
             new MicrodataElementProcessor(),
             new MicrodataContext()
         );